[IMP] tests: Speed up composer tests
Currently, the file `composer.test.ts` takes a lot of time to execute
(43 seconds average). The reason is multi-fold:
1. There are 281 tests
2. Each test instanciate a full spreadsheet, lots of computations that
are not necessary except for ftesting the interactions of the composer
with the spreadsheet component (they currently are strongly linked).

With this revision, the tests covering the interactions with other
components are moved to a dedicated test file, which allows to test the
composer specific behaviour with a more minimalistic setup.

Note that
Benchmark
---------

`test -- composer.test.ts`
before: 43 seconds
after:  9 seconds

`test -- components`
before: 55 seconds
after: 40 seconds

Task 3163569
