+++ src/components/spreadsheet/spreadsheet.ts
@@ -125,8 +125,8 @@ interface SidePanelState {
 }
 
 interface ComposerState {
-  topBarFocus: "inactive" | "contentFocus";
-  gridFocusMode: "inactive" | "cellFocus" | "contentFocus";
+  topBarFocus: Exclude<ComposerFocusType, "cellFocus">;
+  gridFocusMode: ComposerFocusType;
 }
 
 export class Spreadsheet extends Component<SpreadsheetProps, SpreadsheetChildEnv> {
